// src/app/models/empleados.ts
export class Empleados{
    "id": string;
    "nombre": string; 
    "edad": number; 
    "puesto": string; 
    "numero": number; 
    "direccion": string; 
}


export class EmpleadosWithoutID {
    "nombre": string; 
    "edad": number; 
    "puesto": string; 
    "numero": number; 
    "direccion": string; 
}